refactor: moves SamplingUtils to SampleManager
- uses FileManager
